Safari-611.1.10/LayoutTests/fast/workers/constructor-proto-expected.txt

This tests that setting the proto of a built in constructor is not reset

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


proto = new Number(42): 42
Object.getPrototypeOf(WorkerLocation): function () {
    [native code]
}
WorkerLocation.__proto__ = proto: 42
Object.getPrototypeOf(WorkerLocation): 42
PASS successfullyParsed is true

TEST COMPLETE
